pyproject.toml
[project]
name = "supply-manage"
version = "0.1.0"
description = ""
authors = [
    {name = "olshanskiyvv",email = "90987935+olshanskiyvv@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "aiokafka (>=0.12.0,<0.13.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)",
    "greenlet (>=3.2.1,<4.0.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
